+++ metacine/metacine/src/main/java/com/cine/metacine/services/TipoProductoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cine.metacine.models.TipoProducto;
 
-public interface TipoProductoRepository extends JpaRepository<TipoProducto, Integer>{
+public interface TipoProductoRepository extends JpaRepository<TipoProducto, Long>{
 
 }
